server/src/chat/requirements.txt
anyio==4.0.0
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
brotlipy==0.7.0
build==1.0.3
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
Django==4.2.7
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docopt==0.6.2
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jsonpointer==2.1
msgpack==1.0.7
packaging==23.2
phonenumbers==8.13.20
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyproject_hooks==1.0.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarg==0.1.9
zope.interface==6.1
